backend/requirements.txt
python-dotenv
peewee
fastapi
uvicorn[standard]
python-docx
peewee-migrate
psycopg2-binary
